Movie Search works
Ok so the default combo box index is set to -1 which I thought would pull the year 1970. It appears that it actually still points to a NoneType like before but the index is still set to -1(had very strange issues trying to check if NoneType was active). In my date checking searching code, I have it set to skip checking year if the index is -1